Splitting heroku container into a separate image
Dockerfile builds a development version of the app, without an application.

Dockerfile.heroku builds from this image to add the app and change some other settings.

Originally this was going to be a multi-stage Dockerfile but having looked at what this is realised it’s not what we want since we want to keep both images.