Improving CI to only use services when needed
Using docker as a service (for Docker in Docker) requires specific runners. Jobs related to DND use this runner through job level tags but other jobs won’t necessarily because they run normally.

This presents a problem as when a normal job runs the DND service will still be setup (since it’s global) which causes a permisisons problem because that job isn’t ran on a DND enabled runner.

The fix for this is to move the service to the specific job(s) that needs it.