Removing local Snyk image
Snk are still working on host their own image so I have hosted a version based off their repository for now.

This at least means the image isn’t part of this project, where other projects would either need to build their own image, or tie themselves to this project. It also means it can be migrated to an ‘official’ image when ready.