Configuring handled errors to still go to Sentry
Errors such as the file being too large, are handled safely by the application so don’t trigger events in Sentry.

In our case we still want to know when these occur, but we don’t need them reported as actual errors.

In the old Sentry SDK we enabled log reporting, such that logging to the logger would trigger events in Sentry as well. In the new version this isn’t an option so we need to explicitly send these events manually.