Adding request ID support
Currently not used in logging or errors but will be returned in responses allowing others to do something with it.

If a header is already set on the request its value is checked to see if there’s more than one value. A single value is assumed to be non-unique (as a client may have set it) so an additional header value is added. Where this has already happened we assume that one will be unique and leave it.

This implementation will change once changes have been made to the API load balancer so we can tell if it set the request ID or not (using a known prefix). This means we can only add the second request ID when it is more likely to have been set by the client (currently we can’t tell).

This implementation is also likely be come part of a standalone plugin for use in other Flask APIs.