Experimenting with review apps behaviour
Need to test whether the ‘start review’ task is ran everytime a pipeline is ran or not.

If it is we need to add logic to only create a new environment if it doesn’t already exist.

There will need to be work more work on this issue as GitLab doesn’t allow the use of the short branch name in the environment configuration. This means there’s a mismatch between this name and the name passed to Heroku (which is shortened to their max length).