Experimenting with review apps behaviour
GitLab will run the start review app on each commit and since we need to build a Docker image for each commit, creating an application per commit is likely the best way to approach this.

Annoyingly the SHA1 commit exposed is also too long for Heroku so we have to improvise with some other variables that will sufficently unique. It means the environment names will be meaningless, but at least they will exist.